Move copyButton handling to a higher level component
